Examples: Reorganise into applications (not run as part of the CI) and examples (run as part of the CI)
