+++ src/test/java/graphics/scenery/tests/applications/vdi/VDIClientExample.kt
@@ -1,4 +1,4 @@
-package graphics.scenery.tests.examples.volumes
+package graphics.scenery.tests.applications.vdi
 
 import graphics.scenery.SceneryBase
 import graphics.scenery.*
